requirements.txt
requests==2.28.1
Django==4.2
django-filter==23.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
psycopg2==2.9.6
drf-spectacular==0.25.0
